Remove OPTIND use
OPTIND can not reliably store more information than the argument index.
But getopts also needs to store the intra-argument index for options
like `-ab`. Therefore, OPTIND can not be used alone. Up to this commit,
the solution was to use an additional variable. But this still did not work
like the normal getopts, because it could not discern between being
somewhere in the first argument and being restarted.

So from now on, it uses a different variable, which stores the position
like `index.subindex`. This changes the interface from the POSIX
getopts. Drop-in compatibility is thus lost. The changes required to
migrate from POSIX getopts to the one developed here are, however,
limited to a variable rename. And drop-in compatibility has a lower
priority than correctness and unambiguity, which are both maintained.