Add ability to use all printable ASCII characters
Mostly for options like `-?` for help. I don't think actually using
most of the range is a good idea, but if I'm making it compatible with
special chars I might as well catch all of them.

The colon is not supported, as it has a special meaning. Parentheses,
however, are, as long as the opening `(` is the first option in the
optstring.