Remove globmatch, fix zsh compatibility
The globmatch function had a case statement which tried to expand a
variable, and then let the shell interpret the result as a glob. This
does not work on zsh by default. I favor portable code over special cases,
and found no way to get zsh in line without resorting to a setopt call. So
now the actual matching code is hardcoded in the worker function itself,
and globmatch is unnecessary. Zsh almost fully works.